Debian: check if in vbox before installing guest additions

+++ templates/Debian-7.0-i386-netboot/cleanup-virtualbox.sh
@@ -1,4 +1,6 @@
+if test -f .vbox_version ; then
 # Cleanup Virtualbox
 VBOX_VERSION=$(cat .vbox_version)
 VBOX_ISO=VBoxGuestAdditions_$VBOX_VERSION.iso
 rm $VBOX_ISO
+fi
